fix x64 build breaks

+++ src/priv/gfx/util/gpu/ringbuffer.cpp
@@ -81,7 +81,7 @@ GN::gfx::RingBuffer::beginProduce( size_t size )
 
     while( true )
     {
-        int freeSize = mReadPtr - mWritePtr;
+        intptr_t freeSize = mReadPtr - mWritePtr;
         if( freeSize <= 0 ) freeSize += mSize;
         GN_ASSERT( freeSize > 0 );
         freeSize -= 1;
@@ -130,7 +130,7 @@ GN::gfx::RingBuffer::beginConsume( size_t size )
 
     while( true )
     {
-        int unconsumedSize = mWritePtr - mReadPtr;
+        intptr_t unconsumedSize = mWritePtr - mReadPtr;
         if( unconsumedSize < 0 ) unconsumedSize += mSize;
         GN_ASSERT( unconsumedSize >= 0 );
